improve ColorSchemeProvider + use Tailwind hover for Pressable
