+++ packages/design-system/pressable/index.tsx
@@ -1,16 +1,10 @@
 import { ComponentProps, useMemo } from "react";
+import { Pressable as ReactNativePressable } from "react-native";
 
-import { MotiPressable } from "moti/interactions";
-
-import { useIsDarkMode, useOnHover } from "@showtime-xyz/universal.hooks";
-import {
-  styled,
-  colors,
-  tw as tailwind,
-} from "@showtime-xyz/universal.tailwind";
+import { styled } from "@showtime-xyz/universal.tailwind";
 import type { TW } from "@showtime-xyz/universal.tailwind";
 
-export type Props = ComponentProps<typeof MotiPressable> & {
+export type Props = ComponentProps<typeof ReactNativePressable> & {
   tw?: TW;
   /**
    * **DESKTOP WEB_ONLY**
@@ -20,38 +14,17 @@ export type Props = ComponentProps<typeof MotiPressable> & {
   disableHoverEffect?: boolean;
 };
 
-const StyledMotiPressable = styled(MotiPressable);
+const StyledPressable = styled(ReactNativePressable);
 
-export function Pressable({
-  tw,
-  style,
-  disableHoverEffect = false,
-  ...props
-}: Props) {
-  const { onHoverIn, onHoverOut, hovered } = useOnHover();
-  const isDark = useIsDarkMode();
-
-  const hoverStyle = useMemo(() => {
-    if (disableHoverEffect) {
-      return { backgroundColor: "transparent" };
-    }
-    return {
-      backgroundColor: hovered.value
-        ? isDark
-          ? colors.gray[800]
-          : colors.gray[200]
-        : "transparent",
-      transitionDuration: "150ms",
-    };
-  }, [disableHoverEffect, hovered.value, isDark]);
-
-  return (
-    <MotiPressable
-      onHoverIn={onHoverIn}
-      onHoverOut={onHoverOut}
-      style={[hoverStyle, style, tailwind.style(tw)]} // TODO: don't use `tailwind.style`
-      {...props}
-      // tw={Array.isArray(tw) ? tw.join(" ") : tw}
-    />
+export function Pressable({ tw, disableHoverEffect = false, ...props }: Props) {
+  const twWithHover = useMemo(
+    () =>
+      [
+        disableHoverEffect ? "" : "hover:bg-gray-200 dark:hover:bg-gray-800",
+        Array.isArray(tw) ? tw.join(" ") : tw,
+      ].join(" "),
+    [tw, disableHoverEffect]
   );
+
+  return <StyledPressable {...props} tw={twWithHover} />;
 }
